🚑 u is not for float

+++ src/08_Area-Trapecio/main.go
@@ -3,9 +3,9 @@ package main
 import "fmt"
 
 func main() {
-	const baseUno uint8 = 6
-	const baseDos uint8 = 15
-	const alturaTrapecio uint8 = 25
+	const baseUno uint16 = 6
+	const baseDos uint16 = 15
+	const alturaTrapecio uint16 = 25
 
 	const areaTrapecio uint16 = ((baseUno + baseDos) * alturaTrapecio) / 2
